@@ -8,17 +8,18 @@
     "@tanstack/react-query": "^4.36.1",
     "@tanstack/react-table": "^8.8.5",
     "@tanstack/react-virtual": "^3.0.0-beta.54",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
+    "@types/react": "^18.0.20",
+    "@types/react-dom": "^18.0.6",
     "css-loader": "^6.7.3",
     "style-loader": "^3.3.2"
   },
   "scripts": {
     "compile": "gulp compile-extension:positron-data-viewer-ui",
-    "watch": "gulp watch-extension:positron-data-viewer-ui"
+    "watch": "gulp watch-extension:positron-data-viewer-ui",
+    "bundle-dev": "npx webpack-cli watch --mode development --config ./extension.webpack.config.js --output-path ./out"
   }
 }
